Move tests to test_s3_lambda_integration Other notifications are tested there as well
